requirements.txt
numpy
pandas
psycopg2-binary
sqlalchemy
pyarrow
pylint
black
python-dotenv
boto3
pymysql
apache-airflow-providers-amazon